refactor(search): extract backfill code into submodule
- extracts relevant tests
- take opensearch-mock out of search/__tests__ director
- remove testPathIgnorePatterns from jest.config.js
